+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## v1.9.0
+
+*2024-05-24*
+
+- Magic links (RFC, PMID and ISBN) are now parsed as [`MagicLinkToken`](https://github.com/bhsd-harry/wikiparser-node/wiki/MagicLinkToken)
+
 ## v1.8.0
 
 *2024-05-20*